Move some things into `ui._types`
