+++ disnake/ui/select/mentionable.py
@@ -18,7 +18,8 @@ from ...member import Member
 from ...role import Role
 from ...user import ClientUser, User
 from ...utils import MISSING
-from ..item import DecoratedItem, ItemCallbackType, P, V_co, V_deco
+from .._types import P, V_co, V_deco
+from ..item import DecoratedItem, ItemCallbackType
 from .base import BaseSelect, SelectDefaultValueMultiInputType, _create_decorator
 
 if TYPE_CHECKING:
