+++ disnake/ui/select/role.py
@@ -11,18 +11,19 @@ from typing import (
     overload,
 )
 
-from typing_extensions import Self
-
 from ...abc import Snowflake
 from ...components import RoleSelectMenu
 from ...enums import ComponentType, SelectDefaultValueType
 from ...object import Object
 from ...role import Role
 from ...utils import MISSING
-from ..item import DecoratedItem, ItemCallbackType, P, V_co, V_deco
+from .._types import P, V_co, V_deco
+from ..item import DecoratedItem, ItemCallbackType
 from .base import BaseSelect, SelectDefaultValueInputType, _create_decorator
 
 if TYPE_CHECKING:
+    from typing_extensions import Self
+
     from ..view import View
 
 __all__ = (
